listFiles can return null if a different thread removes the dir concurrently (may happen in Solr tests),..
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1598500 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/test-framework/src/java/org/apache/lucene/util/TestUtil.java
@@ -120,14 +120,16 @@ public final class TestUtil {
   }
 
   private static LinkedHashSet<File> rm(LinkedHashSet<File> unremoved, File... locations) {
-    for (File location : locations) {
-      if (location != null && location.exists()) {
-        if (location.isDirectory()) {
-          rm(unremoved, location.listFiles());
-        }
-
-        if (!location.delete()) {
-          unremoved.add(location);
+    if (locations != null) {
+      for (File location : locations) {
+        if (location != null && location.exists()) {
+          if (location.isDirectory()) {
+            rm(unremoved, location.listFiles());
+          }
+  
+          if (!location.delete()) {
+            unremoved.add(location);
+          }
         }
       }
     }
